Add error data to messages in turnstone queue (#876)
This is needed to change Pigeon behavior.  Currently, when a Pigeon
encounters an error relaying a message, the pigeon just attests the
message and then won't touch it again.  This means for a bad message,
the majority of pigeons must try to relay it in order to achieve enough
attestations.

With this new functionality, when an Pigeon encounters an error, it can
send that error data back to Paloma.  That way, all other pigeons can
attest to the error without having to attempt to relay it themselves.

Co-authored-by: Tyler Ruppert <{ID}+{username}@users.noreply.github.com>