Periodically update xchain reference blocks (#1175)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/1637

# Background

At startup, pigeon validates that it's talking to the right chain by
using the reference block height and hash. These values are introduced
when first integrating a chain. However, as time passes, some RPC
providers will prune block information, which makes it impossible for
pigeon to verify the blocks.

To solve this, we introduce a new periodic task that at each 10000
blocks, sends a message requesting pigeons to update the reference block
information (height and hash) for each supported chain.

We add a new queue - `evm/<CHAIN>/reference-block` - and follow the same
flow from ValidatorBalances.

Related pigeons changes are handled in
https://github.com/palomachain/pigeon/pull/393

# Testing completed

- [x] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.