fix: decrease success rate if validator fails to relay (#1295)
# Related Github tickets

- Closes https://github.com/VolumeFi/paloma/issues/2173

# Background

Currently if the assigned validator fails to relay a message it is
simply removed from the queue and does not count as a failure to the
assigned validator. This means if a validator is incapable or unwilling
to relay messages, it's success rate metric never changes.

When removing messages from the queue, we check if it was relayed (it
has public or error data). If it does not, then we consider it a failure
and update the validator metrics.

# Testing completed

- [x] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.